+++ flang/lib/Optimizer/OpenACC/Support/CMakeLists.txt
@@ -1,0 +1,28 @@
+get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
+
+add_flang_library(FIROpenACCSupport
+  FIROpenACCAttributes.cpp
+  FIROpenACCTypeInterfaces.cpp
+  RegisterOpenACCExtensions.cpp
+
+  DEPENDS
+  FIRBuilder
+  FIRDialect
+  FIRDialectSupport
+  FIRSupport
+  HLFIRDialect
+
+  LINK_LIBS
+  FIRBuilder
+  FIRCodeGenDialect
+  FIRDialect
+  FIRDialectSupport
+  FIRSupport
+  HLFIRDialect
+
+  MLIR_DEPS
+  MLIROpenACCDialect
+
+  MLIR_LIBS
+  MLIROpenACCDialect
+)
